pdf2word/requirements.txt
python-docx>=0.8.11
pdf2image>=1.16.0
reportlab>=3.6.12
PyPDF2>=3.0.0
Pillow>=9.0.0
python-dotenv>=0.19.0
tqdm>=4.65.0
colorama>=0.4.6
pdfminer.six>=20221105
logging-utils>=0.3.0 